NL-0003 Add Introduction to DigiD in Netherlands documentation
Added a new file `Introduction.md` under the `Netherlands/Digital-Id` directory to document the DigiD system, which is used as a digital identity solution by the Dutch government. The section describes DigiD's purpose of facilitating secure access to various government services through a registration and login process akin to Single Sign-On (SSO). It outlines the steps for first-time registration, including personal information submission, the receipt of an approval code via mail, and finalizing the registration online using this code. This addition aims to provide a clear understanding of the DigiD system's functionality for users accessing Dutch governmental digital services.